[SHARE-74] Migration of /account/sessions to /session.
