+++ server/index.js
@@ -88,6 +88,7 @@ require('./config/passport')(passport);
 const authRoutes = require('./routes/auth');
 const userRoutes = require('./routes/user');
 const accountRoutes = require('./routes/account');
+const sessionRoutes = require('./routes/session');
 const APIKeyRoutes = require('./routes/apiKey');
 const uploadRoutes = require('./routes/upload');
 const statsRoutes = require('./routes/stats');
@@ -96,6 +97,7 @@ const thirdPartyUploadRoutes = require('./routes/3rd-party/upload');
 app.use('/auth', authRoutes);
 app.use('/user', userRoutes);
 app.use('/account', accountRoutes);
+app.use('/session', sessionRoutes);
 app.use('/apikey', APIKeyRoutes);
 app.use('/upload', uploadRoutes);
 app.use('/stats', statsRoutes);
